Move debounced end_async to the Session class
