+++ plugin/code_actions.py
@@ -6,13 +6,13 @@ from .core.registry import LSPViewEventListener
 from .core.registry import sessions_for_view
 from .core.registry import windows
 from .core.settings import settings
+from .core.types import debounced
 from .core.typing import Any, List, Dict, Callable, Optional, Union, Tuple, Mapping, TypedDict
 from .core.views import entire_content_range
 from .core.views import make_link
 from .core.views import offset_to_point
 from .core.views import region_to_range
 from .core.views import text_document_code_action_params
-from .core.windows import debounced
 from .diagnostics import filter_by_point, view_diagnostics
 from .diagnostics import filter_by_range
 import sublime
