+++ plugin/core/types.py
@@ -1,6 +1,6 @@
 from .collections import DottedDict
 from .logging import debug
-from .typing import Optional, List, Dict, Any, Protocol, Generator
+from .typing import Optional, List, Dict, Any, Protocol, Generator, Callable
 import contextlib
 import sublime
 import time
@@ -12,6 +12,26 @@ def runtime(token: str) -> Generator[None, None, None]:
     t = time.time()
     yield
     debug(token, "running time:", int((time.time() - t) * 1000000), "μs")
+
+
+def debounced(f: Callable[[], None], timeout_ms: int = 0, condition: Callable[[], bool] = lambda: True,
+              async_thread: bool = False) -> None:
+    """
+    Possibly run a function at a later point in time, either on the async thread or on the main thread.
+
+    :param      f:             The function to possibly run
+    :param      timeout_ms:    The time in milliseconds after which to possibly to run the function
+    :param      condition:     The condition that must evaluate to True in order to run the funtion
+    :param      async_thread:  If true, run the function on the async worker thread, otherwise run the function on the
+                               main thread
+    """
+
+    def run() -> None:
+        if condition():
+            f()
+
+    runner = sublime.set_timeout_async if async_thread else sublime.set_timeout
+    runner(run, timeout_ms)
 
 
 class Settings(object):
